toggle functionality
allow toggling to show empty rows and whether to count equipment in containers
